Preference Form
Updated the preference form ui adding some more functionality and finishing what wasn't added. Can be accessed through button on profile page. Can't get it to go to it after signing up for the first time. Submit button uploads the data to Firebase but sometimes the application crashes due to trying to load an activity even though it should be a fragment which I don't know how to fix. Still need to get it to where once data has been stored if you go the preferences page it loads your data into the feilds