Made explore page functional and tweaked Preference Form
Added search bar and a filter. Functionality: any thing typed in the search bar will get compared to users name and bio and if they match they are put into the recycler view. For filter you select checkboxes and then hit submit which will then filter only those who contain those attributes the switch up top makes it so the user has to have all of them or just one. Functionality need: the clear filter button and clear search text button. Preference form, made it so if a user has data already in the firebase then that will be reflected in the preference form