Added algorithm to suggestedFragment and ranking system to Preference page
The recycler view populates but I have not fully tested it. Need to create more users to test it and show it off more. UI is not made at all either. Preference page 2 ui not finished and not all buttons have functionality. Proprpably need to add a function to stop them from submitting when they havent filled out all of the feilds.